RDKB-17280 : syscfg stuck and database corruption.
Reason for change: syscfg stuck and database corruption.

Test Procedure:
Need complete regression testing.

Risks: Low

+++ source/scripts/init/system/utopia_init.sh
@@ -197,6 +197,7 @@ FACTORY_RESET_RGWIFI=y
 FACTORY_RESET_WIFI=w
 SYSCFG_MOUNT=/nvram
 SYSCFG_FILE=$SYSCFG_MOUNT/syscfg.db
+SYSCFG_BKUP_FILE=$SYSCFG_MOUNT/syscfg_bkup.db
 PSM_CUR_XML_CONFIG_FILE_NAME="$SYSCFG_MOUNT/bbhm_cur_cfg.xml"
 PSM_BAK_XML_CONFIG_FILE_NAME="$SYSCFG_MOUNT/bbhm_bak_cfg.xml"
 PSM_TMP_XML_CONFIG_FILE_NAME="$SYSCFG_MOUNT/bbhm_tmp_cfg.xml"
@@ -251,7 +252,13 @@ if [ -f $SYSCFG_FILE ]; then
 	   CheckAndReCreateDB
    fi
 else
-   echo -n > $SYSCFG_FILE
+
+    if [ -f $SYSCFG_BKUP_FILE ]; then 
+	 echo "utopia_init:syscfg.db file is missing, copying backup file to syscfg.db"
+ 	  cp $SYSCFG_BKUP_FILE $SYSCFG_FILE
+    else
+   	   echo -n > $SYSCFG_FILE
+    fi
    syscfg_create -f $SYSCFG_FILE
    if [ $? != 0 ]; then
 	   CheckAndReCreateDB
@@ -326,6 +333,7 @@ fi
    FACTORY_RESET_REASON=true 
    rm -f /nvram/partners_defaults.json 
    rm -f $SYSCFG_FILE
+   rm -f $SYSCFG_BKUP_FILE
    rm -f $PSM_CUR_XML_CONFIG_FILE_NAME
    rm -f $PSM_BAK_XML_CONFIG_FILE_NAME
    rm -f $PSM_TMP_XML_CONFIG_FILE_NAME
